go through bagel language model
